Expose kevent/inotify errnos on thrown error
The error will have a numeric 'errno', which is the system errno
value. (NOT the uv errno value, which is an unrelated value in Node 0.10
and is the negative of the posix errno in Node 0.12.)

In Node 0.12, it will also have a string 'code' like 'ENOENT'. It is too
difficult to get this string with the pre-1.x libuv API.