Add 2220 and 4309 to warning list
Building on win32 is failing for electron version 11.4.7. We need to add
the errors on the warning list to have a passing build on https://github.com/atom/atom/pull/22687

An altenative would be to fix the warnings. Right now there is no
capacity for us to do the fixes